Refactor w/list instead of table
