+++ src/_colors.scss
@@ -1,0 +1,5 @@
+$primary: #82203E;
+$secondary:#DF3461;
+$tertiary: #35183C;
+$white: #FFFFFF;
+$peach: #FC8F8A;
