+++ src/components/App.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from 'react'
-import NavFooter from './layout/NavFooter.jsx'
-import NavHeader from './layout/NavHeader.jsx'
-import Projects from './complex/Projects.jsx'
+import NavFooter from './Navbar/NavFooter.jsx'
+import NavHeader from './Navbar/NavHeader.jsx'
+import Projects from './Projects/Projects.jsx'
 
 export default class App extends Component {
 	render() {
