+++ src/components/Buttons/Button.scss
@@ -1,0 +1,15 @@
+@import "./src/_colors";
+
+.Button {
+	background-color: $white;
+	border: none;
+	color: $peach;
+	cursor: pointer;
+	font-size: 1.2rem;
+	padding: 0;
+
+	i:hover {
+		color: $primary;
+		animation: spin 1s linear 1;
+	}
+}
