+++ src/components/FIles/AddFiles.scss
@@ -1,0 +1,48 @@
+@import "./src/_colors";
+
+.Add-files {
+	.Create-file {
+		display: inline-block;
+		padding: 1rem;
+
+		ol li {
+			padding: 0.35rem;
+		}
+	}
+
+	input[type="text"],
+	input[type="file"] {
+		border: 1px solid $primary;
+		border-radius: 2rem;
+		color: $tertiary;
+		font-size: 16px;
+		height: 2rem;
+		padding: 0 2rem;
+
+		&:focus {
+			outline: none;
+		}
+	}
+
+	input[type="file"] {
+		background-color: $white;
+		height: auto;
+		padding: 6px 5px 3px 5px;
+	}
+
+	.Add-button,
+	.Upload-button {
+		background-color: $white;
+		border: none;
+		color: $peach;
+		cursor: pointer;
+		font-size: 1.2rem;
+		margin-left: -2rem;
+		padding: 0;
+
+		i:hover {
+			color: $primary;
+			animation: spin 1s linear 1;
+		}
+	}
+}
