+++ src/components/FIles/CreateFolder.jsx
@@ -1,12 +1,12 @@
 import React, {Component} from 'react'
-import Button from '../basic/Button.jsx'
-import Input from '../basic/Input.jsx'
-import Row from '../layout/Row.jsx'
+import Button from '../Buttons/Button.jsx'
+import Input from '../Inputs/Input.jsx'
+import Row from '../Row/Row.jsx'
 
 export default class CreateFolder extends Component {
 	render() {
 		return(
-			<Row className="Create-folder">
+			<Row rowClass="Create-folder">
 				<h3>Add a Folder</h3>
 			  <Input
 				  changeHandler={this.props.nameFolder}
@@ -15,9 +15,9 @@ export default class CreateFolder extends Component {
 				  type="text"
 				/>
 		  	<Button
-		  	  className="Add-button"
+		  	  buttonClass="Add-button"
 		  	  clickHandler={this.props.createFolder}
-		  	  iconName="plus"
+		  	  iconClass="plus"
 		  	/>
 			</Row>
 		)
