+++ src/components/Icons/Icon.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Icon({iconClass}) {
+	return (
+		<i className={'fa fa-' + iconClass}></i>
+	)
+}
