+++ src/components/Loaders/BlockLoader.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import Spinner from 'react-spinkit'
-import Row from '../layout/Row.jsx'
+import Row from '../Row/Row.jsx'
 
-export default function Loading({}) {
+export default function BlockLoader({}) {
 	return(
-		<Row className="Loading">
+		<Row rowClass="BlockLoader">
 		  <Spinner name="wave" color="#FC8F8A" />
 		</Row>
 	)
