+++ src/components/Loaders/Loaders.scss
@@ -1,0 +1,6 @@
+.Loading {
+	display: flex;
+	flex-direction: row;
+	justify-content: space-around;
+	height: 6rem;
+}
