+++ src/components/Navbar/NavHeader.jsx
@@ -1,11 +1,12 @@
 import React, {Component} from 'react'
-import Link from '../basic/Link.jsx'
-import Row from './Row.jsx'
+import Link from '../Links/Link.jsx'
+import Row from '../Row/Row.jsx'
+import './Navbar.scss'
 
 export default class NavHeader extends Component {
 	render() {
 		return (
-			<Row className="Header">
+			<Row rowClass="Header">
 				<ul className="Navbar-items">
 					<li>
 						<Link
