+++ src/components/Navbar/Navbar.scss
@@ -1,0 +1,18 @@
+.Navbar-items {
+  display: flex;
+  flex-flow: row nowrap;
+  font-family: 'Quicksand';
+  justify-content: space-between;
+	padding: 1.25rem 0.5rem;
+
+	ul li {
+	  display: inline-block;
+		padding: 0 1rem;
+	}
+
+	a {
+		display: flex;
+		flex-flow: row wrap;
+		justify-content: space-around;
+	}
+}
