+++ src/components/Projects/Projects.scss
@@ -1,0 +1,14 @@
+@import "./src/_colors";
+
+.Projects {
+	text-align: center;
+
+	.Get-list-items {
+		background-color: $white;
+		border: 1px solid $primary;
+		border-radius: 1rem;
+		box-shadow: 0 1px 1px $secondary;
+		text-align: center;
+		vertical-align: middle;
+	}
+}
