+++ src/components/Row/Row.scss
@@ -1,0 +1,4 @@
+.Row {
+	flex: 1 100%;
+	margin-top: 1rem;
+}
