+++ src/styles.scss
@@ -1,204 +1,10 @@
-$primary: #82203E;
-$secondary:#DF3461;
-$tertiary: #35183C;
-$white: #FFFFFF;
-$peach: #FC8F8A;
-
-* {
-  box-sizing: border-box;
-}
-
-html, body {
-	font-family: 'Lato', sans-serif;
-	height: 100%;
-	padding: 0 2%;
-}
-
-body {
-	background-color: $tertiary;
-	display: flex;
-	flex-direction: column;
-}
-
-a {
-	color: $peach;
-	cursor: pointer;
-	text-decoration: none;
-
-	&:hover {
-		color: $white;
-		animation: pulse 1s linear infinite;
-	}
-}
-
-h1, h2, h3, h4, p, ol li {
-	color: $white;
-	font-weight: 300;
-  text-align: center;
-}
-
-ul {
-	padding: 0;
-
-	li {
-		display: inline-block;
-		list-style: none;
-		padding: 1rem;
-	}
-}
+@import "./_colors";
+@import "./_animations";
+@import "./_globals";
 
 .App {
 	.Wrapper {
 		flex: 1 0 auto;
 		height: 100vh;
-	}
-
-	.Row {
-		flex: 1 100%;
-		margin-top: 1rem;
-	}
-
-	.Loading {
-		display: flex;
-		flex-direction: row;
-		justify-content: space-around;
-		height: 6rem;
-	}
-
-	.Navbar-items {
-	  display: flex;
-	  flex-flow: row wrap;
-	  font-family: 'Quicksand';
-	  justify-content: space-between;
-		padding: 1.25rem 0.5rem;
-
-		ul li {
-		  display: inline-block;
-			padding: 0 1rem;
-		}
-
-		a {
-			display: flex;
-			flex-flow: row wrap;
-			justify-content: space-around;
-		}
-	}
-
-	.Projects {
-		text-align: center;
-
-		.Add-files {
-			.Create-file {
-				display: inline-block;
-				padding: 1rem;
-
-				ol li {
-					padding: 0.35rem;
-				}
-			}
-
-			input[type="text"],
-			input[type="file"] {
-				border: 1px solid $primary;
-				border-radius: 2rem;
-				color: $tertiary;
-				font-size: 16px;
-				height: 2rem;
-				padding: 0 2rem;
-
-				&:focus {
-					outline: none;
-				}
-			}
-
-			input[type="file"] {
-				background-color: $white;
-				height: auto;
-				padding: 6px 5px 3px 5px;
-			}
-
-			.Add-button,
-			.Upload-button {
-				background-color: $white;
-				border: none;
-				color: $peach;
-				cursor: pointer;
-				font-size: 1.2rem;
-				margin-left: -2rem;
-				padding: 0;
-
-				i:hover {
-					color: $primary;
-					animation: spin 1s linear 1;
-				}
-			}
-		}
-	}
-
-	.Table {
-		background-color: $white;
-		border: 1px solid $primary;
-		border-radius: 1rem;
-		box-shadow: 0 1px 1px $secondary;
-		text-align: center;
-		vertical-align: middle;
-
-		table {
-			width: 100%;
-
-			.Table-row {
-				&:hover {
-					border-radius: 0.5rem;
-					box-shadow: 0 1px 10px 3px $secondary,
-											0 3px 10px 3px $peach;
-				}
-			}
-
-			.Table-header,
-			.Table-row-cell {
-				padding: 0.5rem;
-			}
-
-			.date,
-			.size,
-			.type {
-				width: 4rem;
-			}
-
-			.name {
-				width: 6rem;
-			}
-		}
-	}
-
-	.Edit-button {
-		background: $white;
-		border: none;
-		cursor: pointer;
-		padding-left: 1rem;
-
-		.fa {
-			color: $secondary;
-			font-size: 1rem;
-
-			&:hover {
-				color: $primary;
-				animation: spin 1s linear 1;
-			}
-		}
-	}
-
-	.hidden {
-		visibility: hidden;
-	}
-
-	// ANIMATION KEYFRAMES
-	@keyframes pulse {
-	  25% {color: $peach;}
-		75% {color: $secondary;}
-	}
-
-	@keyframes spin {
-		100% {transform: rotate(360deg);}
 	}
 }
